refactor: move the multiple file upload test to test_kwargs
