+++ tests/unit/test_datastructures/test_upload_file.py
@@ -1,8 +1,6 @@
 from os import urandom
 from pathlib import Path
-from typing import List, Optional
-
-import pytest
+from typing import Optional
 
 from litestar import post
 from litestar.datastructures import UploadFile
@@ -38,22 +36,6 @@ async def test_upload_file_methods() -> None:
 
     await upload_file.close()
     assert upload_file.file.closed
-
-
-@pytest.mark.parametrize("file_count", (1, 2))
-def test_upload_multiple_files(file_count: int) -> None:
-    @post("/")
-    async def handler(data: List[UploadFile] = Body(media_type=RequestEncodingType.MULTI_PART)) -> None:
-        assert len(data) == file_count
-
-        for file in data:
-            assert await file.read() == b"1"
-
-    with create_test_client([handler]) as client:
-        files_to_upload = [("file", b"1") for _ in range(file_count)]
-        response = client.post("/", files=files_to_upload)
-
-        assert response.status_code == HTTP_201_CREATED
 
 
 def test_cleanup_is_being_performed(tmpdir: Path) -> None:
