+++ tests/unit/test_kwargs/test_multipart_data.py
@@ -394,6 +394,22 @@ def test_image_upload() -> None:
         assert response.status_code == HTTP_201_CREATED
 
 
+@pytest.mark.parametrize("file_count", (1, 2))
+def test_upload_multiple_files(file_count: int) -> None:
+    @post("/")
+    async def handler(data: List[UploadFile] = Body(media_type=RequestEncodingType.MULTI_PART)) -> None:
+        assert len(data) == file_count
+
+        for file in data:
+            assert await file.read() == b"1"
+
+    with create_test_client([handler]) as client:
+        files_to_upload = [("file", b"1") for _ in range(file_count)]
+        response = client.post("/", files=files_to_upload)
+
+        assert response.status_code == HTTP_201_CREATED
+
+
 def test_optional_formdata() -> None:
     @post("/", signature_types=[UploadFile])
     async def hello_world(data: Optional[UploadFile] = Body(media_type=RequestEncodingType.MULTI_PART)) -> None:
